fix ruff formatting error
